% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{.addToCache}
\alias{.addToCache}
\title{Add values to cache}
\usage{
.addToCache(bfc, result, hash)
}
\arguments{
\item{bfc}{Object of class BiocFileCache, created by a call to
BiocFileCache::BiocFileCache()}

\item{result}{character; name of the file written to chache}

\item{hash}{unique hash representing a query.}
}
\value{
Invisibly, \code{TRUE} if the value was added to the cache and
\code{FALSE} if an entry with the same hash already existed.
}
\description{
Add values to cache
}
\keyword{internal}
