% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unexported.R
\name{.cache.writable}
\alias{.cache.writable}
\title{Unexported functions
Test if a path exists and is writable}
\usage{
.cache.writable(path)
}
\arguments{
\item{path}{(\code{character}). The path to be tested.}
}
\value{
TRUE if both conditions are met, FALSE if not.
}
\description{
\command{.cache.writable()} uses \command{file.access()} to test if a
given location exists and is writable by the user.
}
\examples{
\dontrun{.cache.writable(rappdirs::user_cache_dir())}
}
\seealso{
\code{\link[base]{file.access}}
}
\keyword{internal}
