% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{.checkInCache}
\alias{.checkInCache}
\title{Check whether value in cache exists}
\usage{
.checkInCache(bfc, hash, verbose = FALSE)
}
\arguments{
\item{bfc}{Object of class BiocFileCache, created by a call to
BiocFileCache::BiocFileCache()}

\item{hash}{unique hash representing a query.}

\item{verbose}{logical; should additional verbose output be printed? Not currently used.}
}
\value{
\code{TRUE} if a record with the requested hash exists in the file
cache, otherwise \code{FALSE}.
}
\description{
Check whether value in cache exists
}
\details{
This function returns TRUE if a record with the requested hash already
exists in the file cache, otherwise returns FALSE.
}
\keyword{internal}
