% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unexported.R
\name{.chunked_getBM}
\alias{.chunked_getBM}
\title{Make chunked getBM query}
\usage{
.chunked_getBM(
  mart,
  values,
  biom.attributes,
  biom.filter,
  use.cache,
  chunk.size,
  verbose
)
}
\arguments{
\item{mart}{Biomart \code{mart} object.}

\item{values}{\code{character} vector of ids to be converted.}

\item{biom.attributes}{\code{character} vector. Biomart attributes, i.e., type of desired result(s); make sure query id type is included!}

\item{biom.filter}{\code{character} of length one. Name of biomart filter, i.e., type of query ids, defaults to "ensembl_gene_id".}

\item{use.cache}{(\code{logical}). Should \command{getBM()} use the cache? Defaults to \code{TRUE} as in the \command{getBM()} function and is passed on to that.}

\item{chunk.size}{\code{integer} of length one. Maximum number of IDs per BioMart query.
Large ID lists are split into chunks of this size to avoid server timeouts.
Set to \code{Inf} to disable chunking. Defaults to \code{500}.}

\item{verbose}{(\code{logical}). Should verbose output be written to the console?}
}
\value{
A data frame with the retrieved information.
}
\description{
Make chunked getBM query
}
\keyword{internal}
