% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unexported.R
\name{.connect_mart}
\alias{.connect_mart}
\title{Connect to a mart on a given host}
\usage{
.connect_mart(h, biom, biom.data.set, use.cache, verbose)
}
\arguments{
\item{h}{(\code{character}) Host URL.}

\item{biom}{\code{character} vector. Biomart to use (uses the first element of the vector), defaults to "ensembl".}

\item{biom.data.set}{\code{character} of length one. Biomart data set to use. Defaults to 'human' (internally translated to "hsapiens_gene_ensembl" if \code{biom.mart="ensembl"}).}

\item{use.cache}{(\code{logical}) Should \emph{\code{biomaRt}} functions use the cache? Defaults to \code{TRUE}.}

\item{verbose}{Logical.}
}
\value{
A \code{mart} object for use in biomart queries.
}
\description{
Connect to a mart on a given host
}
\keyword{internal}
