% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unexported.R
\name{.get.Ensembl_config}
\alias{.get.Ensembl_config}
\title{Unexported functions
Test and retrieve Ensembl-specific CURL SSL configuration.}
\usage{
.get.Ensembl_config(use.cache = TRUE)
}
\arguments{
\item{use.cache}{(\code{logical}) Should \emph{\code{biomaRt}} functions use the cache? Defaults to \code{TRUE}.}
}
\value{
A R object of class \code{request} listing current CURL options.
}
\description{
\command{.get.Ensembl_config()} tests and gets CURL options used with \emph{"^https://.*ensembl.org"} URLs.
The function is a modified version of \code{.getEnsemblSSL} from the \emph{\code{biomaRt}} package.
}
\examples{
\dontrun{.get.Ensembl_config()}
}
\seealso{
\code{\link[httr]{config}}
}
\keyword{internal}
