% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{.readFromCache}
\alias{.readFromCache}
\title{Read values from cache}
\usage{
.readFromCache(bfc, hash)
}
\arguments{
\item{bfc}{Object of class BiocFileCache, created by a call to
BiocFileCache::BiocFileCache()}

\item{hash}{unique hash representing a query.}
}
\value{
The cached R object stored under \code{hash}.
}
\description{
Read values from cache
}
\keyword{internal}
