% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertId2.R
\docType{package}
\name{convertid}
\alias{convertid-package}
\alias{convertid}
\title{Convert Gene IDs Between Each Other and Fetch Annotations from Biomart}
\description{
Gene Symbols, Ensembl Gene IDs or Entrez Gene IDs are converted
    using the Bimap interface in 'AnnotationDbi' in convertId2() for the most
    common use cases in data analysis. The main function in the package is
    convert.bm() which queries Biomart using the full capacity of the API
    provided through the 'biomaRt' package. Presets and defaults are provided
    for convenience but all "marts", "filters" and "attributes" can be set by
    the user. Function convert.alias() converts Gene Symbols to Aliases and
    vice versa and function likely_symbol() attempts to determine the most
    likely current Gene Symbol.
}
\details{
\tabular{ll}{
Package: \tab convertid\cr
Type: \tab Package\cr
Initial version: \tab 0.1-0\cr
Created: \tab 2021-08-18\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Vidal Fey <vidal.fey@gmail.com>
Maintainer: Vidal Fey <vidal.fey@gmail.com>
}
\keyword{internal}
\keyword{package}
