% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unexported.R
\name{.create.cache}
\alias{.create.cache}
\title{Unexported functions
Create and initialise a BioC cache at a given location.}
\usage{
.create.cache(cache.path = rappdirs::user_cache_dir("biomaRt"))
}
\arguments{
\item{cache.path}{(\code{character}). The path to use for the cached files.}
}
\value{
TRUE if the location was successfully set up, FALSE if not.
}
\description{
\command{.create.cache()} uses \command{BiocFileCache::BiocFileCache()} to set up
a BioC file cache at a location defined by the user. See the \command{BiocFileCache} package
for more information.
}
\examples{
\dontrun{.create.cache(rappdirs::user_cache_dir("biomaRt"))}
}
\seealso{
\code{\link[BiocFileCache]{BiocFileCache}}
}
\keyword{internal}
